pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiplex"
version = "0.1.0"
description = "Multiplex Virtual IHC: Brightfield to multi-marker prediction"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Tom Li"}
]
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "scikit-image>=0.21",
    "tifffile>=2023.7",
    "aicsimageio>=4.11",
    "albumentations>=1.3",
    "opencv-python>=4.8",
    "pillow>=10.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "tqdm>=4.65",
    "pyyaml>=6.0",
    "wandb>=0.15",
    "timm>=1.0.0",
    "torchinfo>=1.8.0",
    "einops>=0.7.0",
    "lpips>=0.1.4",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
